refactor: move functions from processmanagement.go into seperate files for clarity
