+++ src/core/processmanagement.go
@@ -3,21 +3,13 @@ package core
 
 import (
 	"StationeersServerUI/src/config"
-	"StationeersServerUI/src/ssestream"
-	"bufio"
 	"fmt"
-	"io"
 	"net/http"
-	"os"
 	"os/exec"
 	"runtime"
-	"strconv"
 	"strings"
 	"sync"
 	"syscall"
-	"time"
-
-	"github.com/fsnotify/fsnotify"
 )
 
 var (
@@ -33,81 +25,6 @@ const (
 	colorCyan   = "\033[36m"
 	colorBold   = "\033[1m"
 )
-
-type Arg struct {
-	Flag          string
-	Value         string
-	RequiresValue bool
-	Condition     func() bool
-	NoQuote       bool
-}
-
-func buildCommandArgs() []string {
-	var argOrder = []Arg{
-		{Flag: "-nographics", RequiresValue: false},
-		{Flag: "-batchmode", RequiresValue: false},
-		{Flag: "-LOAD", Value: config.SaveFileName, RequiresValue: true, NoQuote: true}, // LOAD has special handling because the gameserver expects 2 parameters
-		{Flag: "-logFile", Value: "./debug.log", Condition: func() bool { return runtime.GOOS == "linux" }, RequiresValue: true},
-		{Flag: "-settings", RequiresValue: false},
-		{Flag: "StartLocalHost", Value: strconv.FormatBool(config.StartLocalHost), RequiresValue: true},
-		{Flag: "ServerVisible", Value: strconv.FormatBool(config.ServerVisible), RequiresValue: true},
-		{Flag: "GamePort", Value: config.GamePort, RequiresValue: true},
-		{Flag: "UPNPEnabled", Value: strconv.FormatBool(config.UPNPEnabled), RequiresValue: true},
-		{Flag: "ServerName", Value: config.ServerName, RequiresValue: true},
-		{Flag: "ServerPassword", Value: config.ServerPassword, Condition: func() bool { return config.ServerPassword != "" }, RequiresValue: true},
-		{Flag: "ServerMaxPlayers", Value: config.ServerMaxPlayers, RequiresValue: true},
-		{Flag: "AutoSave", Value: strconv.FormatBool(config.AutoSave), RequiresValue: true},
-		{Flag: "SaveInterval", Value: config.SaveInterval, RequiresValue: true},
-		{Flag: "ServerAuthSecret", Value: config.ServerAuthSecret, Condition: func() bool { return config.ServerAuthSecret != "" }, RequiresValue: true},
-		{Flag: "UpdatePort", Value: config.UpdatePort, RequiresValue: true},
-		{Flag: "AutoPauseServer", Value: strconv.FormatBool(config.AutoPauseServer), RequiresValue: true},
-		{Flag: "UseSteamP2P", Value: strconv.FormatBool(config.UseSteamP2P), RequiresValue: true},
-		{Flag: "AdminPassword", Value: config.AdminPassword, Condition: func() bool { return config.AdminPassword != "" }, RequiresValue: true},
-	}
-
-	var args []string
-	for _, arg := range argOrder {
-		if arg.Condition != nil && !arg.Condition() {
-			continue
-		}
-		if arg.RequiresValue && arg.Value == "" {
-			continue
-		}
-
-		args = append(args, arg.Flag)
-
-		if arg.Flag == "-LOAD" && arg.Value != "" {
-			parts := strings.SplitN(arg.Value, " ", 2)
-			for _, part := range parts {
-				if part != "" {
-					args = append(args, part)
-				}
-			}
-			continue
-		}
-
-		if arg.Value != "" {
-			args = append(args, arg.Value)
-		}
-	}
-
-	if config.AdditionalParams != "" {
-		args = append(args, strings.Fields(config.AdditionalParams)...)
-	}
-
-	if config.LocalIpAddress != "" {
-		args = append(args, "LocalIpAddress")
-		args = append(args, config.LocalIpAddress)
-	}
-
-	if config.IsDebugMode {
-		fmt.Println("=== DEBUG: Raw arguments passed to exec.Command ===")
-		for i, arg := range args {
-			fmt.Printf("Arg[%d]: %q\n", i, arg)
-		}
-	}
-	return args
-}
 
 func StartServer(w http.ResponseWriter, r *http.Request) {
 	mu.Lock()
@@ -170,141 +87,6 @@ func StartServer(w http.ResponseWriter, r *http.Request) {
 	}
 
 	fmt.Fprintf(w, "Server started.")
-}
-
-// readPipe for Windows
-func readPipe(pipe io.ReadCloser) {
-	scanner := bufio.NewScanner(pipe)
-	if config.IsDebugMode {
-		fmt.Println("Started reading pipe") // Debug
-	}
-	for scanner.Scan() {
-		output := scanner.Text()
-		ssestream.BroadcastConsoleOutput(output)
-	}
-	if err := scanner.Err(); err != nil {
-		if config.IsDebugMode {
-			fmt.Println("Pipe error:", err) // Debug
-		}
-		ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error reading pipe: %v", err))
-	}
-	if config.IsDebugMode {
-		fmt.Println("Pipe closed") // Debug
-	}
-}
-
-// tailLogFile implements a tail -f-like behavior using fsnotify
-func tailLogFile(logFilePath string) {
-	// Wait briefly to ensure the file exists after server start
-	time.Sleep(1 * time.Second)
-
-	// Open the log file
-	file, err := os.Open(logFilePath)
-	if err != nil {
-		if config.IsDebugMode {
-			fmt.Printf("Error opening log file: %v\n", err)
-		}
-		ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error opening log file: %v", err))
-		return
-	}
-	defer file.Close()
-
-	// Create an fsnotify watcher
-	watcher, err := fsnotify.NewWatcher()
-	if err != nil {
-		if config.IsDebugMode {
-			fmt.Printf("Error creating watcher: %v\n", err)
-		}
-		ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error spawning log file watcher: %v", err))
-		return
-	}
-	defer watcher.Close()
-
-	// Add the log file to the watcher
-	err = watcher.Add(logFilePath)
-	if err != nil {
-		if config.IsDebugMode {
-			fmt.Printf("Error adding log file to watcher: %v\n", err)
-		}
-		ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error adding log file to watcher: %v", err))
-		return
-	}
-
-	// Seek to the end of the file initially to start tailing from the current point
-	_, err = file.Seek(0, io.SeekEnd)
-	if err != nil {
-		if config.IsDebugMode {
-			fmt.Printf("Error seeking to end of log file: %v\n", err)
-		}
-		ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error seeking log file: %v", err))
-		return
-	}
-
-	scanner := bufio.NewScanner(file)
-	if config.IsDebugMode {
-		fmt.Println("Started tailing log file with fsnotify")
-	}
-
-	// Goroutine to handle file events
-	go func() {
-		for {
-			select {
-			case event, ok := <-watcher.Events:
-				if !ok {
-					return
-				}
-				if event.Op&fsnotify.Write == fsnotify.Write {
-					// New content has been written to the file
-					for scanner.Scan() {
-						output := scanner.Text()
-						if config.IsDebugMode {
-							fmt.Println("DEBUG: Read from log file:", output)
-						}
-						ssestream.BroadcastConsoleOutput(output)
-					}
-					if err := scanner.Err(); err != nil {
-						if config.IsDebugMode {
-							fmt.Printf("Error reading log file: %v\n", err)
-						}
-						ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error reading log file: %v", err))
-						return
-					}
-				}
-				// Handle file truncation or rotation (optional)
-				if event.Op&fsnotify.Remove == fsnotify.Remove || event.Op&fsnotify.Rename == fsnotify.Rename {
-					if config.IsDebugMode {
-						fmt.Println("Log file removed or renamed, attempting to reopen")
-					}
-					// Reopen the file and re-add to watcher
-					file.Close()
-					file, err = os.Open(logFilePath)
-					if err != nil {
-						ssestream.BroadcastConsoleOutput(fmt.Sprintf("Error reopening log file: %v", err))
-						return
-					}
-					scanner = bufio.NewScanner(file)
-					watcher.Add(logFilePath)
-				}
-			case err, ok := <-watcher.Errors:
-				if !ok {
-					return
-				}
-				if config.IsDebugMode {
-					fmt.Printf("Watcher error: %v\n", err)
-				}
-				ssestream.BroadcastConsoleOutput(fmt.Sprintf("Watcher error: %v", err))
-			case <-logDone:
-				return
-			}
-		}
-	}()
-
-	// Keep the function running until the server stops
-	<-logDone
-}
-
-func GetLogOutput(w http.ResponseWriter, r *http.Request) {
-	ssestream.StartConsoleStream()(w, r)
 }
 
 func StopServer(w http.ResponseWriter, r *http.Request) {
